hf/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
apturl==0.5.2
argcomplete==3.1.1
async-timeout==4.0.2
attrs==23.1.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.1.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
contourpy==1.1.0
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
datasets==2.13.1
dbus-python==1.2.18
decorator==5.1.1
defer==1.0.6
dill==0.3.6
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
duplicity==0.8.21
evaluate==0.4.0
fasteners==0.14.1
filelock==3.12.2
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
future==0.18.2
httplib2==0.20.2
huggingface-hub==0.15.1
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.7.1
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.4
multiprocess==0.70.14
netifaces==0.11.0
numpy==1.25.0
oauthlib==3.2.0
olefile==0.46
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.3
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
pipenv==2023.6.26
pipx==1.2.0
platformdirs==3.8.0
protobuf==3.12.4
psutil==5.9.4
ptyprocess==0.7.0
pyarrow==12.0.1
pycairo==1.20.1
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyte==0.8.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
regex==2023.6.3
reportlab==3.6.8
requests==2.25.1
responses==0.18.0
safetensors==0.3.1
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.4.1
systemd-python==234
thefuck==3.32
tokenizers==0.13.3
tqdm==4.65.0
transformers==4.30.2
typing_extensions==4.7.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
userpath==1.8.0
virtualenv==20.23.1
virtualenv-clone==0.5.7
wadllib==1.3.6
wcwidth==0.2.6
xdg==5
xkit==0.0.0
xxhash==3.2.0
yarl==1.9.2
zipp==1.0.0
